Add account suggestion with naive bayesian inference
When asking for an account, a list of suggested accounts will be computed by
running naive bayesian inference over the description, comments and accounts of
the current transaction.  This list is displayed in the context when the user
hasn't typed anything.

To be consistent with the suggestion that already existed in previous versions,
this suggestions is added as the first element of the list.